VFD-116: removed earlier added SuomiFi hack - now each suomifi login has unique user tied to it, one user per one suomifi auth method, per environment.
