+++ VirtualFinland.UserAPI/src/VirtualFinland.UsersAPI/Activities/Identity/IdentityController.cs
@@ -29,7 +29,7 @@ public class IdentityController : ControllerBase
     [ProducesErrorResponseType(typeof(ProblemDetails))]
     public async Task<IActionResult> VerifyIdentityUser()
     {
-        var user = await _mediator.Send(new VerifyIdentityUser.Query(this.User.FindFirst(ClaimTypes.NameIdentifier)?.Value ?? this.User.FindFirst(Constants.Web.ClaimNameId)?.Value, this.User.Claims.First().Issuer));
+        var user = await _mediator.Send(new VerifyIdentityUser.Query(this.User.FindFirst(ClaimTypes.NameIdentifier)?.Value ?? this.User.FindFirst(Constants.Web.ClaimUserId)?.Value, this.User.Claims.First().Issuer));
 
         return Ok(user);
     }
