+++ VirtualFinland.UserAPI/src/VirtualFinland.UsersAPI/Helpers/Services/AuthenticationService.cs
@@ -1,6 +1,3 @@
-using System.Security.Claims;
-using VirtualFinland.UserAPI.Data;
-
 namespace VirtualFinland.UserAPI.Helpers.Services;
 
 public class AuthenticationService
@@ -16,15 +13,9 @@ public class AuthenticationService
         Guid? currentUserId = null;
         if (currentUserId == null)
         {
-            if (httpRequest.Path.Value != null &&
-                httpRequest.HttpContext.User != null &&
-                httpRequest.HttpContext.User.Identity != null &&
-                httpRequest.HttpContext.User.Identity.IsAuthenticated &&
-                !httpRequest.Path.Value.Contains("/identity"))
-            {
-                var dbUser = await _userSecurityService.VerifyAndGetAuthenticatedUser(httpRequest.HttpContext.User.Claims.First().Issuer, httpRequest.HttpContext.User.FindFirst(ClaimTypes.NameIdentifier)?.Value ?? httpRequest.HttpContext.User.FindFirst(Constants.Web.ClaimNameId)?.Value);
-                currentUserId = dbUser?.Id;
-            }
+            var token = httpRequest.Headers.Authorization.ToString().Replace("Bearer ", string.Empty);
+            var dbUser = await _userSecurityService.VerifyAndGetAuthenticatedUser(token);
+            currentUserId = dbUser?.Id;
         }
         return currentUserId;
     }
